have variables functions also return the err_ret remove assignment and err checks in a single line, for readability also do the same for increment and array index, keep each line to do as much as it needs to, for readability sya.c: change iter variable types from uint16_t back to int, rename error to err, for consistency multi.c: check err after sep_str in case malloc fails

+++ multi.h
@@ -94,7 +94,7 @@ err_ret extractValue(char **input, int varIndex, struct vari *var, struct matrix
  * int *tok
  *   int to determine if '-' is negative or minus
  */
-err_ret chk_var(const char *buffer, char *input, uint16_t *iter, struct vari *var, struct stack *num, struct stack *ch);
+err_ret chk_var(const char *buffer, char *input, int *iter, struct vari *var, struct stack *num, struct stack *ch);
 
 /*
  * remove spaces from input
@@ -127,7 +127,7 @@ err_ret printLine(char **input, struct vari *var);
  * int *iterator
  *   the main loop counter in sya, so it can be incremented the appropriate amount
  */
-char **sep_str(char *input, char const * const limiter, char const * const delimiter, uint16_t *iter, err_ret *error);
+char **sep_str(char *input, char const * const limiter, char const * const delimiter, int *iter, err_ret *error);
 
 /*
  * free string output from separateString
